Extract player personna
