+++ packages/client/src/modules/play/PlayerPersona/index.tsx
@@ -1,0 +1,3 @@
+import { PlayerPersona } from "./PlayerPersona";
+
+export { PlayerPersona };
